Unifying parallel executors, initializing from launch policy
