+++ cmake/HPX_SetupBoost.cmake
@@ -113,7 +113,7 @@ hpx_add_config_define(BOOST_PARAMETER_MAX_ARITY 7)
 if(MSVC)
   HPX_option(HPX_WITH_BOOST_ALL_DYNAMIC_LINK BOOL "Add BOOST_ALL_DYN_LINK to compile flags" OFF)
   if (HPX_WITH_BOOST_ALL_DYNAMIC_LINK)
-    hpx_add_compile_flag(-DBOOST_ALL_DYN_LINK)
+    hpx_add_config_define(BOOST_ALL_DYN_LINK)
   endif()
 else()
   hpx_add_config_define(HPX_COROUTINE_NO_SEPARATE_CALL_SITES)
