+++ hpx/parallel/executors.hpp
@@ -10,7 +10,6 @@
 #include <hpx/parallel/executors/timed_executor_traits.hpp>
 #include <hpx/parallel/executors/executor_parameter_traits.hpp>
 #include <hpx/parallel/executors/parallel_executor.hpp>
-#include <hpx/parallel/executors/parallel_fork_executor.hpp>
 #include <hpx/parallel/executors/sequential_executor.hpp>
 #include <hpx/parallel/executors/service_executors.hpp>
 #include <hpx/parallel/executors/thread_pool_executors.hpp>
